requirements2.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
gtts

fastapi
fastapi-sqlalchemy
uvicorn
assemblyai
openai==0.28
Jinja2==3.0.3
requests
gtts
python-multipart
